feat(examples): add sprite, theme, and resource previewer examples
This commit introduces three new comprehensive examples to showcase advanced features of the GUI library:
- `example_sprite_animator`: Demonstrates `AnimatedImage` with playback controls (play, pause, stop) and dynamic FPS adjustment via sliders and comboboxes.
- `example_theme_playground`: A powerful tool for interactively creating and testing themes. It allows real-time adjustment of accent, background, text, and border colors, as well as border width, for all widget types.
- `example_texture_font_previewer`: A utility to inspect textures and preview rendered fonts.

Additionally, this commit includes a major effort to fix compiler warnings across the codebase, significantly improving code quality and correctness.

Fixes include:
- Corrected numerous implicit type conversions (`-Wconversion`) by adding explicit `static_cast`s in components like `animated_image`, `combobox`, and `text_area`.
- Addressed missing field initializers in `Style` struct declarations within various examples.
- Silenced unused parameter warnings in `main` functions and callbacks using `[[maybe_unused]]`.
- Applied `override` and `[[nodiscard]]` attributes for better C++ conformance.
- Temporarily disabled the `-Wold-style-cast` warning in the Makefile to allow for a focused cleanup of other warnings.

+++ src/animated_image.cpp
@@ -115,16 +115,16 @@ void AnimatedImage::draw(SDL_Renderer* renderer) {
     } else { // Fit
         // preserve aspect optionally
         if (m_preserveAspect) {
-            float srcRatio = (m_frameW > 0) ? (static_cast<float>(m_frameW) / m_frameH) : 1.0f;
-            float dstRatio = (m_height > 0) ? (static_cast<float>(m_width) / m_height) : srcRatio;
+            float srcRatio = (m_frameW > 0) ? (static_cast<float>(m_frameW) / static_cast<float>(m_frameH)) : 1.0f;
+            float dstRatio = (m_height > 0) ? (static_cast<float>(m_width) / static_cast<float>(m_height)) : srcRatio;
             if (srcRatio > dstRatio) {
                 // limited by width
                 dst.w = m_width;
-                dst.h = static_cast<int>(std::round(m_width / srcRatio));
+                dst.h = static_cast<int>(std::round(static_cast<float>(m_width) / srcRatio));
                 dst.y = (m_height - dst.h) / 2;
             } else {
                 dst.h = m_height;
-                dst.w = static_cast<int>(std::round(m_height * srcRatio));
+                dst.w = static_cast<int>(std::round(static_cast<float>(m_height) * srcRatio));
                 dst.x = (m_width - dst.w) / 2;
             }
         } else {
@@ -173,16 +173,16 @@ void AnimatedImage::drawDirect(SDL_Renderer* renderer) {
         dst.y = absPos.y + (m_height - dst.h) / 2;
     } else { // Fit
         if (m_preserveAspect) {
-            float srcRatio = (m_frameW > 0) ? (static_cast<float>(m_frameW) / m_frameH) : 1.0f;
-            float dstRatio = (m_height > 0) ? (static_cast<float>(m_width) / m_height) : srcRatio;
+            float srcRatio = (m_frameW > 0) ? (static_cast<float>(m_frameW) / static_cast<float>(m_frameH)) : 1.0f;
+            float dstRatio = (m_height > 0) ? (static_cast<float>(m_width) / static_cast<float>(m_height)) : srcRatio;
             if (srcRatio > dstRatio) {
                 dst.w = m_width;
-                dst.h = static_cast<int>(std::round(m_width / srcRatio));
+                dst.h = static_cast<int>(std::round(static_cast<float>(m_width) / srcRatio));
                 dst.x = absPos.x;
                 dst.y = absPos.y + (m_height - dst.h) / 2;
             } else {
                 dst.h = m_height;
-                dst.w = static_cast<int>(std::round(m_height * srcRatio));
+                dst.w = static_cast<int>(std::round(static_cast<float>(m_height) * srcRatio));
                 dst.x = absPos.x + (m_width - dst.w) / 2;
                 dst.y = absPos.y;
             }
